+++ lib/MPU/Calibration/CalibrationManager.cpp
@@ -214,38 +214,57 @@ void CalibrationManager::calibration()
         this->accelgyro->setZGyroOffset(gz_offset);
 
         this->meansensors();
-        Serial.println("...");
+        Serial.println("...\t");
 
         if (abs(mean_ax) <= acel_deadzone)
+        {
+            Serial.print("X\t");
             ready++;
+        }
         else
             ax_offset = ax_offset - mean_ax / acel_deadzone;
 
         if (abs(mean_ay) <= acel_deadzone)
+        {
+            Serial.print("Y\t");
             ready++;
+        }
         else
             ay_offset = ay_offset - mean_ay / acel_deadzone;
 
-        if (abs(16384 - mean_az) <= acel_deadzone)
+        if (abs(16384 / 2 - mean_az) <= acel_deadzone)
+        {
+            Serial.print("Z\t");
             ready++;
+        }
         else
             az_offset = az_offset + (16384 - mean_az) / acel_deadzone;
 
         if (abs(mean_gx) <= giro_deadzone)
+        {
+            Serial.print("GX\t");
             ready++;
+        }
         else
             gx_offset = gx_offset - mean_gx / (giro_deadzone + 1);
 
         if (abs(mean_gy) <= giro_deadzone)
+        {
+            Serial.print("GY\t");
             ready++;
+        }
         else
             gy_offset = gy_offset - mean_gy / (giro_deadzone + 1);
 
         if (abs(mean_gz) <= giro_deadzone)
+        {
+            Serial.print("GZ\t");
             ready++;
+        }
         else
             gz_offset = gz_offset - mean_gz / (giro_deadzone + 1);
 
+        Serial.print("\n");
         if (ready == 6)
             break;
     }
@@ -257,7 +276,7 @@ void CalibrationManager::calibration()
  */
 int CalibrationManager::needCalibration()
 {
-    if (digitalRead(25))
+    if (digitalRead(15) == false)
     {
         return 1;
     }