+++ lib/debug/debug.h
@@ -1,0 +1,28 @@
+
+#ifndef _DEBUG_
+#define _DEBUG_
+#define DEBUG_MODE true
+#include "Arduino.h"
+#include "helper_3dmath.h"
+
+// void printVector(VectorFloat v)
+// {
+//     Serial.print(v.x);
+//     Serial.print("\t");
+//     Serial.print(v.y);
+//     Serial.print("\t");
+//     Serial.print(v.z);
+//     Serial.println("");
+// }
+
+// void printVector(VectorInt16 v)
+// {
+//     Serial.print(v.x);
+//     Serial.print("\t");
+//     Serial.print(v.y);
+//     Serial.print("\t");
+//     Serial.print(v.z);
+//     Serial.println("");
+// }
+
+#endif